feat: middleware 삭제
