+++ proxy.ts
@@ -2,15 +2,30 @@ import { NextResponse } from 'next/server';
 
 import type { NextRequest } from 'next/server';
 
+const RETURN_PATH_COOKIE = 'sossbar-login-return';
+
 const PROTECTED_PREFIXES = ['/mypage', '/personal', '/projects', '/reviews', '/profile', '/signup'];
 
-/** OG 크롤러·초대 링크 수신자가 로그인 없이 접근해야 하는 경로 */
-const isPublicShareRoute = (pathname: string, searchParams: URLSearchParams): boolean => {
-  if (pathname.startsWith('/profile/')) {
-    return true;
-  }
+const CRAWLER_UA_PATTERNS = [
+  'facebookexternalhit',
+  'twitterbot',
+  'linkedinbot',
+  'slackbot',
+  'telegrambot',
+  'whatsapp',
+  'kakaotalk',
+  'googlebot',
+  'bingbot',
+];
 
-  if (pathname === '/projects' && searchParams.has('inviteProjectId')) {
+const isCrawler = (request: NextRequest): boolean => {
+  const ua = (request.headers.get('user-agent') ?? '').toLowerCase();
+  return CRAWLER_UA_PATTERNS.some((pattern) => ua.includes(pattern));
+};
+
+/** 비인증 접근 시 홈의 로그인 팝업으로 유도해야 하는 공유 링크 경로 */
+const isSharedLinkRoute = (pathname: string, searchParams: URLSearchParams): boolean => {
+  if (pathname.startsWith('/profile/')) {
     return true;
   }
 
@@ -26,11 +41,34 @@ export const proxy = (request: NextRequest) => {
 
   const isProtected = PROTECTED_PREFIXES.some((p) => pathname === p || pathname.startsWith(`${p}/`));
 
-  if (isProtected && !request.cookies.has('accessToken') && !isPublicShareRoute(pathname, searchParams)) {
-    return NextResponse.redirect(new URL('/', request.url));
+  if (!isProtected || request.cookies.has('accessToken')) {
+    return NextResponse.next();
   }
 
-  return NextResponse.next();
+  // 크롤러: OG 메타데이터 수집을 위해 공유 링크 경로는 통과
+  if (isSharedLinkRoute(pathname, searchParams) && isCrawler(request)) {
+    return NextResponse.next();
+  }
+
+  // /projects?inviteProjectId=X: 로그인 없이 접근 허용
+  if (pathname === '/projects' && searchParams.has('inviteProjectId')) {
+    return NextResponse.next();
+  }
+
+  // 공유 링크: 홈 화면 위에 로그인 팝업 표시 + 복귀 경로 저장
+  if (isSharedLinkRoute(pathname, searchParams)) {
+    const query = searchParams.toString();
+    const returnPath = query ? `${pathname}?${query}` : pathname;
+    const response = NextResponse.redirect(new URL('/?modal=login', request.url));
+    response.cookies.set(RETURN_PATH_COOKIE, encodeURIComponent(returnPath), {
+      path: '/',
+      sameSite: 'lax',
+      maxAge: 300,
+    });
+    return response;
+  }
+
+  return NextResponse.redirect(new URL('/', request.url));
 };
 
 export const config = {
